+++ clang-tidy/nodecpp/NakedPtrFromFunctionCheck.cpp
@@ -20,15 +20,16 @@ namespace nodecpp {
 
 void NakedPtrFromFunctionCheck::registerMatchers(MatchFinder *Finder) {
   
-
-
   Finder->addMatcher(callExpr(hasType(pointerType())).bind("call"), this);
 }
 
 void NakedPtrFromFunctionCheck::check(const MatchFinder::MatchResult &Result) {
-  const auto *m = Result.Nodes.getNodeAs<CallExpr>("call");
 
-//  diag(m->getLocStart(), "call found here!", DiagnosticIDs::Note);
+	const auto *m = Result.Nodes.getNodeAs<CallExpr>("call");
+  if (isa<CXXOperatorCallExpr>(m) || isa<CXXMemberCallExpr>(m))
+    return;
+
+      //  diag(m->getLocStart(), "call found here!", DiagnosticIDs::Note);
 
   if (isParentCompStmt(Result.Context, m) || isParentVarDecl(Result.Context, m)) {
     // this is ok