+++ clang-tidy/nodecpp/NodeCppTidyModule.cpp
@@ -15,6 +15,7 @@
 #include "NakedPtrFieldCheck.h"
 #include "NakedPtrFromFunctionCheck.h"
 #include "NakedPtrFromMethodCheck.h"
+#include "NakedPtrFromOperatorCheck.h"
 #include "NakedPtrFuncCheck.h"
 #include "NewExprCheck.h"
 #include "NoCastCheck.h"
@@ -39,6 +40,8 @@ public:
         "nodecpp-naked-ptr-from-function");
     CheckFactories.registerCheck<NakedPtrFromMethodCheck>(
         "nodecpp-naked-ptr-from-method");
+    CheckFactories.registerCheck<NakedPtrFromOperatorCheck>(
+        "nodecpp-naked-ptr-from-operator");
     CheckFactories.registerCheck<NakedPtrFuncCheck>(
         "nodecpp-naked-ptr-func");
     CheckFactories.registerCheck<NewExprCheck>(